refactor(workflow): move get_workflow_run_status to queries.rs and dedupe get_fan_out_items SQL
Addresses two of the seven blocking review findings on PR #2768:

**Architecture (`get_workflow_run_status` location):** the function queries
the `workflow_runs` table, not the fan-out items table, so it belongs in
`queries.rs` alongside `is_run_cancelled` and `get_workflow_run` rather than
in `fan_out.rs`. This was acceptable while the function was a private
method on `WorkflowManager`, but becomes a misleading import path
(`workflow::manager::fan_out::get_workflow_run_status`) once the function
is promoted to a `pub fn` in a `pub(crate)` module. Moves the free function
and its shim from `fan_out.rs` to `queries.rs`.

**DRY (`get_fan_out_items` SQL):** the two query branches differed only by
the `WHERE` clause predicate. Consolidate into a single `format!`-built SQL
with a conditional `:status` clause, threading the status param into the
`Vec<(&str, &dyn ToSql)>` only when the filter is set. Same pattern used
in `list_all_workflow_runs_filtered_paginated` (queries.rs).

Drop the now-unused `OptionalExtension` import from fan_out.rs.

The remaining five test-coverage findings are addressed in the PR
discussion: `approve_gate`, `get_gate_approval_state`,
`validate_gate_selections`, `get_fan_out_items_checked`, and
`skip_fan_out_items_by_item_ids` already have substantial test coverage in
`workflow/manager/tests.rs` and `workflow/tests/gates.rs` — the reviewer
saw "no tests in the diff" but didn't check the existing test files.
